brkn: test  v3.0.1 -> gLit_GUI

+++ 3dLand/GlLand/gLit_GUI/src/ctrls/font.cpp
@@ -35,7 +35,7 @@ Font::Font(const char* _fontPath, const int _fontSize) {
       texture,
       ivec2(ftFace->glyph->bitmap.width, ftFace->glyph->bitmap.rows),
       ivec2(ftFace->glyph->bitmap_left, ftFace->glyph->bitmap_top),
-      (int)ftFace->glyph->advance.x
+      ftFace->glyph->advance.x
     };
